Move shared validation logic to a helper
