+++ src/lib/users/dialogs/inst_user_first.rb
@@ -23,6 +23,7 @@ require "users/dialogs/users_to_import"
 require "y2users"
 require "y2users/users_simple"
 require "y2users/help_texts"
+require "y2users/inst_users_dialog_helper"
 require "users/users_database"
 require "tmpdir"
 
@@ -32,6 +33,8 @@ module Yast
   # will then be created by that module during inst_finish
   # rubocop:disable Metrics/ClassLength
   class InstUserFirstDialog < ::UI::InstallationDialog
+    include Y2Users::InstUsersDialogHelper
+
     # Widgets to enable/disable depending on the selected action
     # (the first one receives the initial focus if applicable)
     WIDGETS = {
@@ -250,39 +253,6 @@ module Yast
       Progress.set(@progress_orig)
     end
 
-    # Config object holding the users and passwords to create
-    #
-    # @return [Y2Users::Config]
-    def users_config
-      return @users_config if @users_config
-
-      @users_config = Y2Users::Config.new
-      Y2Users::UsersSimple::Reader.new.read_to(@users_config)
-      @users_config
-    end
-
-    # All users to be created
-    #
-    # @return [Array<Y2Users::User>]
-    def users
-      users_config.users.reject(&:root?)
-    end
-
-    # User to be created, useful during the :new_user action in which {#users}
-    # is known to contain only one element
-    #
-    # @return [Y2Users::User]
-    def user
-      users.first
-    end
-
-    # Root users for which is possible to define the password during the :new_user action
-    #
-    # @return [Y2Users::User]
-    def root_user
-      users_config.users.find(&:root?)
-    end
-
     # TODO: let's simply return false for the time being
     def imported?(_user)
       false
@@ -441,50 +411,6 @@ module Yast
       UsersSimple.SkipRootPasswordDialog(false)
       UsersSimple.SetRootPassword("") if root_dialog_follows
       UsersSimple.SetAutologinUser("")
-    end
-
-    # Checks whether the information entered for the user is valid, reporting the problem to
-    # the user otherwise
-    #
-    # @return [Boolean]
-    def valid_user?
-      issue = user.issues.first
-      if issue
-        Report.Error(issue.message)
-        focus_on(issue.location)
-        return false
-      end
-
-      true
-    end
-
-    # Checks whether the entered password is acceptable, reporting fatal problems to the user and
-    # asking for confirmation for the non-fatal ones
-    #
-    # @return [Boolean]
-    def valid_password?
-      user_to_validate = @use_pw_for_root ? root_user : user
-      issues = user_to_validate.password_issues
-      return true if issues.empty?
-
-      UI.SetFocus(Id(:pw1))
-
-      fatal = issues.find(&:fatal?)
-      if fatal
-        Report.Error(fatal.message)
-        return false
-      end
-
-      message = issues.map(&:message).join("\n\n") + "\n\n" + _("Really use this password?")
-      Popup.YesNo(message)
-    end
-
-    # Sets the UI focus in the widget corresponding to the given issue location
-    #
-    # @param location [Y2Issues::Location]
-    def focus_on(location)
-      id = (location.path == "name") ? :username : :full_name
-      UI.SetFocus(Id(id))
     end
 
     def dialog_content
@@ -646,6 +572,15 @@ module Yast
 
     def import_available?
       !!@import_available
+    end
+
+    # User for performing password validatons
+    #
+    # @see Y2Users::InstUsersDialogHelper#user_to_validate
+    #
+    # @return [Y2Users::User] root user if using entered password for it too; the new user otherwise
+    def user_to_validate
+      @use_pw_for_root ? root_user : user
     end
   end
   # rubocop:enable Metrics/ClassLength
